使用mapstruct替代beanutil；修复重置密码错误的bug；导入单元测试依赖

+++ ruoyi-demo/src/main/java/com/ruoyi/demo/domain/bo/TestDemoBo.java
@@ -3,17 +3,16 @@ package com.ruoyi.demo.domain.bo;
 import com.ruoyi.common.core.domain.BaseEntity;
 import com.ruoyi.common.core.validate.AddGroup;
 import com.ruoyi.common.core.validate.EditGroup;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
-
 /**
- * 测试单表业务对象 test_demo
+ * 测试单业务对象 test_demo
  *
- * @author Lion Li
- * @date 2021-07-26
+ * @author ruoyi
+ * @date 2024-04-09
  */
 
 @Data
@@ -23,37 +22,38 @@ public class TestDemoBo extends BaseEntity {
     /**
      * 主键
      */
-    @NotNull(message = "主键不能为空", groups = {EditGroup.class})
+    @NotNull(message = "主键不能为空", groups = { EditGroup.class })
     private Long id;
 
     /**
      * 部门id
      */
-    @NotNull(message = "部门id不能为空", groups = {AddGroup.class, EditGroup.class})
+    @NotNull(message = "部门id不能为空", groups = { AddGroup.class, EditGroup.class })
     private Long deptId;
 
     /**
      * 用户id
      */
-    @NotNull(message = "用户id不能为空", groups = {AddGroup.class, EditGroup.class})
+    @NotNull(message = "用户id不能为空", groups = { AddGroup.class, EditGroup.class })
     private Long userId;
 
     /**
      * 排序号
      */
-    @NotNull(message = "排序号不能为空", groups = {AddGroup.class, EditGroup.class})
-    private Integer orderNum;
+    @NotNull(message = "排序号不能为空", groups = { AddGroup.class, EditGroup.class })
+    private Long orderNum;
 
     /**
      * key键
      */
-    @NotBlank(message = "key键不能为空", groups = {AddGroup.class, EditGroup.class})
+    @NotBlank(message = "key键不能为空", groups = { AddGroup.class, EditGroup.class })
     private String testKey;
 
     /**
      * 值
      */
-    @NotBlank(message = "值不能为空", groups = {AddGroup.class, EditGroup.class})
+    @NotBlank(message = "值不能为空", groups = { AddGroup.class, EditGroup.class })
     private String value;
+
 
 }
